requirements.txt
llama-index
llama-index-embeddings-openai
llama-index-vector-stores-mongodb
datasets
pymongo
pandas
llama-index-readers-web
llama-index-embeddings-fireworks
jupyter
streamlit
python-dotenv
